- Fixed wrong 2018.2 timestamp handling - Fixed wrong 2018.2 worklog item filtering

+++ application/src/main/java/de/pbauerochse/worklogviewer/fx/tasks/WorklogViewerTask.kt
@@ -20,8 +20,8 @@ abstract class WorklogViewerTask<T> : Task<T>(), ProgressCallback {
 
     override fun call(): T = start(this)
 
-    override fun setProgress(message: String, amount: Int) {
-        updateProgress(amount.toLong(), 100)
+    override fun setProgress(message: String, amount: Double) {
+        updateProgress(amount, 100.toDouble())
         updateMessage(message)
     }
 }
